requirements/base.txt
PyMySQL==0.6.6
# Postgesql anyone?
# psycopg2
Django==1.8.2
djangorestframework==3.1.2
#djangorestframework-jwt==1.5.0
djangorestframework-jwt==1.2.0
drf-nested-routers==0.9.0
#wsgiref==0.1.2
django-custom-user==0.5

Pillow==2.8.1
django-resized==0.3.5
sorl-thumbnail==12.2

django-redis==4.0.0
hiredis==0.2.0

django-compressor==1.5
